+++ framework/CHANGELOG.md
@@ -74,6 +74,7 @@ Yii Framework 2 Change Log
 - Enh #1973: `yii message/extract` is now able to generate `.po` files (SergeiKutanov, samdark)
 - Enh #1984: ActionFilter will now mark event as handled when action run is aborted (cebe)
 - Enh #2003: Added `filter` property to `ExistValidator` and `UniqueValidator` to support adding additional filtering conditions (qiangxue)
+- Enh #2051: Do not save null data into database when using RBAC (qiangxue)
 - Enh: Added `favicon.ico` and `robots.txt` to default application templates (samdark)
 - Enh: Added `Widget::autoIdPrefix` to support prefixing automatically generated widget IDs (qiangxue)
 - Enh: Support for file aliases in console command 'message' (omnilight)
@@ -84,6 +85,7 @@ Yii Framework 2 Change Log
 - Enh: `init` of advanced application now allows to specify answer for overwriting files via `init --overwrite=n` (samdark)
 - Enh: Added `TableSchema::fullName` property (qiangxue)
 - Enh #1839: Added support for getting file extension and basename from uploaded file (anfrantic)
+- Enh: yii\codeception\TestCase now supports loading and using fixtures via Yii fixture framework (qiangxue)
 - Chg #1519: `yii\web\User::loginRequired()` now returns the `Response` object instead of exiting the application (qiangxue)
 - Chg #1586: `QueryBuilder::buildLikeCondition()` will now escape special characters and use percentage characters by default (qiangxue)
 - Chg #1610: `Html::activeCheckboxList()` and `Html::activeRadioList()` will submit an empty string if no checkbox/radio is selected (qiangxue)
@@ -106,7 +108,10 @@ Yii Framework 2 Change Log
 - New #66: [Auth client library](https://github.com/yiisoft/yii2-authclient) OpenId, OAuth1, OAuth2 clients (klimov-paul)
 - New #1393: [Codeception testing framework integration](https://github.com/yiisoft/yii2-codeception) (Ragazzo)
 - New #1438: [MongoDB integration](https://github.com/yiisoft/yii2-mongodb) ActiveRecord and Query (klimov-paul)
+- New #1956: Implemented test fixture framework (qiangxue)
 - New: Yii framework now comes with core messages in multiple languages
+- New: Added yii\codeception\DbTestCase (qiangxue)
+
 
 2.0.0 alpha, December 1, 2013
 ---------------------------